persiste  y es un priblema
